started loading small dataset to local cache

+++ include/psdd/fpga_evaluate.h
@@ -7,5 +7,5 @@ double EvaluateWithoutPointer(const std::bitset<MAX_VAR> &variables,
                       uint32_t fpga_serialized_psdd_evaluate [PSDD_SIZE],
                       FPGAPsddNodeStruct fpga_node_vector[PSDD_SIZE],
                       uint32_t children_vector[TOTAL_CHILDREN],
-                      double parameter_vector [TOTAL_PARAM]);
+                      float parameter_vector [TOTAL_PARAM]);
 #endif // FPGA_STRUCTURED_BAYESIAN_NETWORK_EVALUATE_H
